# api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Dossier API",
    "description" : "The Dossier sync + tracking contract. Implement this spec to build a Dossier-compatible backend the extension and web app can point at. Authenticate at POST /api/authenticate, then send the access token as a Bearer token on every protected request; refresh it at POST /api/refresh.",
    "contact" : { },
    "license" : { },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/bios/{id}" : {
      "get" : {
        "tags" : [ "bio-resource" ],
        "operationId" : "getBio",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BioDTO"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "bio-resource" ],
        "operationId" : "updateBio",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BioDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BioDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "bio-resource" ],
        "operationId" : "deleteBio",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "bio-resource" ],
        "operationId" : "partialUpdateBio",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BioDTO"
              }
            },
            "application/merge-patch+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BioDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BioDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bios" : {
      "get" : {
        "tags" : [ "bio-resource" ],
        "operationId" : "getAllBios",
        "parameters" : [ {
          "name" : "eagerload",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BioDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "bio-resource" ],
        "operationId" : "createBio",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BioDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BioDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/authorities" : {
      "get" : {
        "tags" : [ "authority-resource" ],
        "operationId" : "getAllAuthorities",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Authority"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "authority-resource" ],
        "operationId" : "createAuthority",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Authority"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Authority"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/authorities/{id}" : {
      "get" : {
        "tags" : [ "authority-resource" ],
        "operationId" : "getAuthority",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Authority"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "authority-resource" ],
        "operationId" : "deleteAuthority",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "login" : {
            "type" : "string"
          }
        }
      },
...
      "BioDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "payload" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserDTO"
          }
        },
        "required" : [ "updatedAt" ]
      },
...
      "Authority" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 0
          }
        },
        "required" : [ "name" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "description" : "JWT access token issued by POST /api/authenticate (or refreshed via POST /api/refresh).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}